move rxdata files into directory DrugCompare
