+++ app/app-routing.module.ts
@@ -5,7 +5,7 @@ import { ChecklistComponent } from './checklist/checklist.component';
 import { CaseManagementComponent } from './CaseManagement/casemanagement.component';
 
 import { HomeComponent } from './home.component';
-import { RXDataComponent } from './rx-data.component';
+import { RXDataComponent } from './DrugCompare/rx-data.component';
 import { ValuePropComponent } from './value-prop.component';
 import { LineGraphComponent } from './line-graph.component';
