+++ app/app.module.ts
@@ -1,7 +1,7 @@
 import { NgModule }      from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
-import { RxDataService } from './rx-data.service';
-import { RxDataFilter }  from './rx-data-filter.component';
+import { RxDataService } from './DrugCompare/rx-data.service';
+import { RxDataFilter }  from './DrugCompare/rx-data-filter.component';
 
 import { FormsModule } from "@angular/forms";
 import { HttpModule, JsonpModule  }    from '@angular/http';
@@ -13,8 +13,8 @@ import { AppComponent }  from './app.component';
 import { CaseManagementComponent } from './CaseManagement/casemanagement.component';
 import { CaseManagementService } from './CaseManagement/casemanagement.service';
 import { HomeComponent } from './home.component';
-import { RXDataComponent } from './rx-data.component';
-import { RxDataFilterPipe }      from './rx-data-filter.pipe';
+import { RXDataComponent } from './DrugCompare/rx-data.component';
+import { RxDataFilterPipe }      from './DrugCompare/rx-data-filter.pipe';
 import { ValuePropComponent } from './value-prop.component';
 import { ChecklistComponent } from './checklist/checklist.component';
 import { PatientSummaryComponent } from './checklist/patient-summary.component';
